Started adding test cases

+++ src/main/java/com/poc/cache/cachedemo/controllers/EmployeeController.java
@@ -2,52 +2,75 @@ package com.poc.cache.cachedemo.controllers;
 
 import com.poc.cache.cachedemo.models.Employee;
 import com.poc.cache.cachedemo.repositories.EmployeeRepository;
+import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
 
 @RestController
 @RequestMapping(path = "/api/v1")
+@AllArgsConstructor
 public class EmployeeController {
 
-    @Autowired
     private EmployeeRepository employeeRepository;
 
-    @PostMapping("/employee")
+    @PostMapping(value = "/employee", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Employee> createEmployee(@RequestBody Employee employee) {
         Employee savedEmployee = employeeRepository.save(employee);
-        return new ResponseEntity(savedEmployee, HttpStatus.CREATED);
+        return new ResponseEntity<>(savedEmployee, HttpStatus.CREATED);
     }
 
+    //TODO: get employee from cache if non existent call h2 database and cache the same
 
-    //get employee from cache if non existent call h2 database and cache the same
-
-    //get all employee
-
-    @PutMapping("/employee/{id}")
-    public ResponseEntity<Employee> updateEmployee(@RequestParam("id") String id, @RequestBody Employee employee) {
+    @GetMapping(value = "/employee/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<?> getEmployee(@PathVariable("id") String id) {
         Employee savedEmployee = employeeRepository.findById(Long.parseLong(id)).orElse(null);
-//TODO: missing null case
+        if (savedEmployee == null) {
+            return new ResponseEntity<>("Employee not found", HttpStatus.OK);
+        }
+        return new ResponseEntity<>(savedEmployee, HttpStatus.OK);
+    }
+
+    @GetMapping(value = "/employee", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<List<Employee>> getAllEmployee() {
+        List<Employee> employeeList = employeeRepository.findAll();
+        return ResponseEntity.ok(employeeList);
+    }
+
+    @PutMapping(value = "/employee/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<?> updateEmployee(@PathVariable("id") String id, @RequestBody Employee employee) {
+        Employee savedEmployee = employeeRepository.findById(Long.parseLong(id)).orElse(null);
+        if (savedEmployee == null) {
+            return new ResponseEntity<>("Employee not found", HttpStatus.OK);
+        }
         savedEmployee.setName(employee.getName());
         savedEmployee.setRole(employee.getRole());
         savedEmployee.setPhoneNumber(employee.getPhoneNumber());
 
         Employee returnedEmployee = employeeRepository.save(savedEmployee);
 
-        return new ResponseEntity(returnedEmployee, HttpStatus.NO_CONTENT);
+        return new ResponseEntity<>(returnedEmployee, HttpStatus.OK);
     }
 
-    @DeleteMapping("/employee/{id}")
-    public ResponseEntity<String> deleteEmployee(@RequestParam("id") String id) {
+    @DeleteMapping(value = "/employee/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<String> deleteEmployee(@PathVariable("id") String id) {
+        Employee savedEmployee = employeeRepository.findById(Long.parseLong(id)).orElse(null);
+        if (savedEmployee == null) {
+            return new ResponseEntity<>("Employee not found", HttpStatus.OK);
+        }
+
         employeeRepository.deleteById(Long.parseLong(id));
-        return new ResponseEntity("employee deleted", HttpStatus.OK);
-        //TODO: what if id not present
+        return new ResponseEntity<>("Employee deleted", HttpStatus.OK);
     }
 }
